Fixed folder within folder creation
When we moved the full paths out of the javascript, we fogot to make
directory creation relative, and instead returned only the name of the
new folder to the creation script, causing it to be created at the top
level. Fixed by updating the javascript call to include all
subdirectories needed.